// src/global/ts/enums.ts
/* Bootstrap Enums */

export enum ButtonType {
  submit = 'submit',
  button = 'button',
  reset = 'reset',
}

export enum Color {
  white = 'white',
  grayscale01 = 'grayscale01',
  grayscale02 = 'grayscale02',
  grayscale03 = 'grayscale03',
  grayscale04 = 'grayscale04',
  grayscale05 = 'grayscale05',
  grayscale06 = 'grayscale06',
  grayscale07 = 'grayscale07',
  black = 'black',
  primary01 = 'primary01',
  primary02 = 'primary02',
  secondary01 = 'secondary01',
  accent01 = 'accent01',
  hover01 = 'hover01',
  alertgreen = 'alertgreen',
  tabletint = 'tabletint',
}

export enum FormCheckboxType {
  checkbox = 'checkbox',
  radio = 'radio',
}

export enum FormTextInputType {
  email = 'email',
  password = 'password',
  text = 'text',
  number = 'number',
  search = 'search',
}

export enum GridType {
  container = 'container',
  containerfluid = 'container-fluid',
  row = 'row',
  col = 'col',
}

export enum Justify {
  center = 'center',
  start = 'start',
  end = 'end',
}

export enum LogoColor {
  white = 'white',
  black = 'black',
  blue = 'blue',
}

export enum Size {
  sm = 'sm',
  lg = 'lg',
}

export enum SlideDirection {
  right = 'right',
  left = 'left',
}

/* Arvest Enums */

export enum AccordionType {
  table = 'table',
  tableSimple = 'table-simple',
  content = 'content',
  nav = 'nav',
  mobileNav = 'mobileNav'
}

export enum BadgeColor {
  label = 'label',
  counter = 'counter'
}

export enum CardSelector {
  localFavorites = 'Local Favorites',
  aToZ = 'A-Z',
  zToA = 'Z-A'
}

export enum CellVariant {
  singleText,
  bulletList,
  checkmark,
  dash,
  nonBulletList,
  na,
}

export enum StepperStates {
  active = 'active',
  complete = 'complete',
  incomplete = 'incomplete'
}

export enum TypographyVariants {
  h1 = 'h1',
  h2 = 'h2',
  h3 = 'h3',
  h4 = 'h4',
  subheading = 'subheading',
  body1 = 'body1',
  body2 = 'body2',
  action = 'action',
  navigation = 'navigation',
  mobileNavigation = 'mobileNavigation',
  labels = 'labels',
  caption = 'caption',
}

export enum TooltipPosition {
  topLeft = 'top-left',
  topCenter = 'top-center',
  topRight = 'top-right',
  bottomLeft = 'bottom-left',
  bottomCenter = 'bottom-center',
  bottomRight = 'bottom-right'
}

export enum TooltipCta {
  text = 'text',
  button = 'button'
}

export enum TooltipTrigger {
  hover = 'hover',
  click = 'click'
}

export enum ViewportBreakpoints {
  mobile = 375,
  tablet = 768,
  desktopSmall = 1241,
  desktop = 1440,
  primaryHeroBreakpoint = 1020,
}

export enum SubNavTypes {
  heading = 'heading',
  link = 'link',
  button = 'button',
  ad = 'ad'
}

export enum MountStates {
  UNMOUNTED = 'unmounted',
  MOUNTING = 'mounting',
  MOUNTED = 'mounted',
  UNMOUNTING = 'unmounting',
}